cleanup from review
