+++ near-sdk/src/lib.rs
@@ -10,7 +10,7 @@ pub use near_sdk_macros::{
 pub mod store;
 
 #[cfg(feature = "unstable")]
-pub use environment::hash;
+pub use environment::hash as crypto_hash;
 
 pub mod collections;
 mod environment;
