+++ near-sdk/src/utils/stable_map.rs
@@ -1,4 +1,4 @@
-use core::borrow::Borrow;
+use std::borrow::Borrow;
 use std::cell::RefCell;
 use std::collections::BTreeMap;
